refactor: 폴더 구조 정리 및 주석 정리
